Use extension function to zip list into pairs
Thanks to the hint from Kotlin team at JetBrains, there's an extension
function in the standard library that can zip the list of elements in
the list, and it is possible to apply a predicate to this function.
It returns a list of booleans. Then just count those that are `true`.